Rate functionality working, began work on individual item info (with ASCII art!)

+++ mediastore/src/mediastore/Album.java
@@ -22,7 +22,7 @@ public class Album extends Media {
      * @param price price of the album
      * @param numSold number of albums sold (initialized to 0)
      */
-    public Album( int id, String author, String title, int duration, String genre, int rating, int totalReviews, double price, int numSold ) {
+    public Album( int id, String author, String title, int duration, String genre, double rating, int totalReviews, double price, int numSold ) {
 
         super( id, author, title, duration, genre, rating, totalReviews, price, numSold );
